Rename directory to components
