+++ testing/features/support/components/asset_hyperlink.rb
