+++ testing/features/support/components/oisc_warning/sticky.rb
